+++ tourplanner.Models/Tour.cs
@@ -5,6 +5,6 @@
         public int tour_ID { get; set; }
         public string tour_Name { get; set; }
         public string tour_Description { get; set; }
-        public float tour_Distance { get; set; }
+        public double tour_Distance { get; set; }
     }
 }
